+++ src/test/e2e/commands/RMB/rmb-status.test.ts
@@ -19,7 +19,10 @@ suite("status: Right Click Menu", function () {
     configurationName = await LiquibaseGUITestUtils.setupTests();
   });
 
-  LiquibaseGUITestUtils.createRmbArguments("Status...", ContextOptions.NO_CONTEXT).forEach((pArgument) => {
+  LiquibaseGUITestUtils.createRmbArguments(
+    "List the not deployed changesets (status)...",
+    ContextOptions.NO_CONTEXT
+  ).forEach((pArgument) => {
     /**
      * Test case for executing the 'status' command from RMB.
      */